pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helix"
version = "0.1.0"
description = "Computation-aware execution optimizer for LLM agent workflows."
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["click>=8.1", "pyyaml>=6.0", "rich>=13.0"]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.25"]
dev = ["pytest>=7.0", "pytest-cov", "mypy", "ruff"]
all = ["helix[openai,anthropic,dev]"]

[project.scripts]
helix = "helix.cli.main:cli"

[tool.setuptools.packages.find]
include = ["helix*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
ignore_errors = true
ignore_missing_imports = true
